+++ src/main/java/com/moaplace/mapper/RentalMapper.java
@@ -13,5 +13,5 @@ public interface RentalMapper {
 	List<RentalVO> list(HashMap<String, Object> map);
 	int updateState(HashMap<String, Object> map);
 	List<RentalCalendarDTO> getSchedules(HashMap<String, String> map);
-	
+	RentalVO detail(int rental_num);
 }
