+++ src/main/java/com/moaplace/vo/RentalVO.java
@@ -2,6 +2,8 @@ package com.moaplace.vo;
 
 import java.sql.Date;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -18,6 +20,7 @@ public class RentalVO {
 	private String rental_email;
 	private String rental_title;
 	private String rental_genre;
+	@JsonFormat(shape=JsonFormat.Shape.STRING, pattern="yyyy-MM-dd")
 	private Date rental_date;
 	private String rental_time;
 	private String rental_savefilename;
@@ -27,6 +30,7 @@ public class RentalVO {
 	private String rental_ownsphone;
 	private String rental_ownemail;
 	private String rental_content;
+	@JsonFormat(shape=JsonFormat.Shape.STRING, pattern="yyyy-MM-dd")
 	private Date regdate;
 	private String rental_state;
 }
